# examples/blog/internal/transport/http/blog.go
// Package http provides HTTP handlers for the service.
// ============================================================================
// AUTO-GENERATED by ANG (Architectural Normalized Generator)
// Version: 0.1.127
// InputHash: 2280e6dff5bcac137d5e4ca79079af522793ed3c7f1301a7955910011cd5b7e5
// CompilerHash: 294b50c17c5868adf98ff6e7e34d9b0ddcf67156d863f2f508ad57df44889164
// DO NOT EDIT - changes will be overwritten on next generation
// GENERATED: true
// ============================================================================

package http

import (
	"crypto/sha256"
	"encoding/json"
	"fmt"
	"net/http"
	"strconv"

	"github.com/example/blog/internal/pkg/errors"
	"github.com/example/blog/internal/port"
	"github.com/go-chi/chi/v5"
)

func RegisterBlogRoutes(r chi.Router, svc port.Blog) {
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware).Delete("/comments/{id}", handlerDeleteComment(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware).Put("/comments/{id}", handlerUpdateComment(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware, RequireRoles([]string{"author", "admin"})).Get("/my/posts", handlerListMyPosts(svc))
	r.With(MaxBodySizeMiddleware(1048576)).Get("/posts", handlerListPosts(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware, RequireRoles([]string{"author", "admin"})).Post("/posts", handlerCreatePost(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware, RequireRoles([]string{"author", "admin"})).Delete("/posts/{id}", handlerDeletePost(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware, RequireRoles([]string{"author", "admin"})).Put("/posts/{id}", handlerUpdatePost(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware, RequireRoles([]string{"author", "admin"})).Post("/posts/{id}/archive", handlerArchivePost(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware, RequireRoles([]string{"admin"})).Post("/posts/{id}/publish", handlerPublishPost(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware, RequireRoles([]string{"author", "admin"})).Post("/posts/{id}/submit", handlerSubmitPost(svc))
	r.With(MaxBodySizeMiddleware(1048576)).Get("/posts/{postID}/comments", handlerListComments(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware).Post("/posts/{postID}/comments", handlerCreateComment(svc))
	r.With(MaxBodySizeMiddleware(1048576)).Get("/posts/{slug}", handlerGetPost(svc))
	r.With(MaxBodySizeMiddleware(1048576)).Get("/tags", handlerListTags(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware, RequireRoles([]string{"admin"})).Post("/tags", handlerCreateTag(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware, RequireRoles([]string{"admin"})).Delete("/tags/{id}", handlerDeleteTag(svc))
	r.With(MaxBodySizeMiddleware(1048576), AuthMiddleware, RequireRoles([]string{"admin"})).Put("/tags/{id}", handlerUpdateTag(svc))
}

// handlerDeleteComment
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerDeleteComment(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.DeleteCommentRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if v := chi.URLParam(r, "id"); v != "" {
			req.ID = v
		}
		if req.UserID == "" {
			req.UserID = CurrentUserID(r)
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.DeleteComment(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerUpdateComment
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerUpdateComment(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.UpdateCommentRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if v := chi.URLParam(r, "id"); v != "" {
			req.ID = v
		}
		if req.UserID == "" {
			req.UserID = CurrentUserID(r)
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.UpdateComment(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerListMyPosts
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerListMyPosts(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.ListMyPostsRequest
		if v := r.URL.Query().Get("userid"); v != "" {
			req.UserID = v
		}
		if v := r.URL.Query().Get("status"); v != "" {
			req.Status = v
		}
		if v := r.URL.Query().Get("limit"); v != "" {
			if i, err := strconv.Atoi(v); err == nil {
				req.Limit = i
			} else {
				errors.WriteError(w, r, errors.NewValidationError("Invalid query", map[string]string{"limit": "must be integer"}))
				return
			}
		}
		if v := r.URL.Query().Get("offset"); v != "" {
			if i, err := strconv.Atoi(v); err == nil {
				req.Offset = i
			} else {
				errors.WriteError(w, r, errors.NewValidationError("Invalid query", map[string]string{"offset": "must be integer"}))
				return
			}
		}
		if req.UserID == "" {
			req.UserID = CurrentUserID(r)
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.ListMyPosts(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		var out any
		out = resp
		data, err := json.Marshal(out)
		if err != nil {
			errors.WriteError(w, r, errors.New(http.StatusInternalServerError, "Internal Server Error", "Failed to encode response"))
			return
		}
		etag := fmt.Sprintf("W/\"%x\"", sha256.Sum256(data))
		w.Header().Set("ETag", etag)
		if match := r.Header.Get("If-None-Match"); match == etag {
			w.WriteHeader(http.StatusNotModified)
			return
		}
		_, _ = w.Write(data)

	}
}

// handlerListPosts
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerListPosts(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.ListPostsRequest
		if v := r.URL.Query().Get("tag"); v != "" {
			req.Tag = v
		}
		if v := r.URL.Query().Get("limit"); v != "" {
			if i, err := strconv.Atoi(v); err == nil {
				req.Limit = i
			} else {
				errors.WriteError(w, r, errors.NewValidationError("Invalid query", map[string]string{"limit": "must be integer"}))
				return
			}
		}
		if v := r.URL.Query().Get("offset"); v != "" {
			if i, err := strconv.Atoi(v); err == nil {
				req.Offset = i
			} else {
				errors.WriteError(w, r, errors.NewValidationError("Invalid query", map[string]string{"offset": "must be integer"}))
				return
			}
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.ListPosts(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		var out any
		out = resp
		data, err := json.Marshal(out)
		if err != nil {
			errors.WriteError(w, r, errors.New(http.StatusInternalServerError, "Internal Server Error", "Failed to encode response"))
			return
		}
		etag := fmt.Sprintf("W/\"%x\"", sha256.Sum256(data))
		w.Header().Set("ETag", etag)
		if match := r.Header.Get("If-None-Match"); match == etag {
			w.WriteHeader(http.StatusNotModified)
			return
		}
		_, _ = w.Write(data)

	}
}

// handlerCreatePost
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerCreatePost(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.CreatePostRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if req.UserID == "" {
			req.UserID = CurrentUserID(r)
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.CreatePost(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerDeletePost
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerDeletePost(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.DeletePostRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if v := chi.URLParam(r, "id"); v != "" {
			req.ID = v
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.DeletePost(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerUpdatePost
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerUpdatePost(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.UpdatePostRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if v := chi.URLParam(r, "id"); v != "" {
			req.ID = v
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.UpdatePost(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerArchivePost
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerArchivePost(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.ArchivePostRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if v := chi.URLParam(r, "id"); v != "" {
			req.ID = v
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.ArchivePost(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerPublishPost
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerPublishPost(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.PublishPostRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if v := chi.URLParam(r, "id"); v != "" {
			req.ID = v
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.PublishPost(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerSubmitPost
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerSubmitPost(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.SubmitPostRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if v := chi.URLParam(r, "id"); v != "" {
			req.ID = v
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.SubmitPost(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerListComments
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerListComments(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.ListCommentsRequest
		if v := r.URL.Query().Get("postid"); v != "" {
			req.PostID = v
		}
		if v := r.URL.Query().Get("limit"); v != "" {
			if i, err := strconv.Atoi(v); err == nil {
				req.Limit = i
			} else {
				errors.WriteError(w, r, errors.NewValidationError("Invalid query", map[string]string{"limit": "must be integer"}))
				return
			}
		}
		if v := r.URL.Query().Get("offset"); v != "" {
			if i, err := strconv.Atoi(v); err == nil {
				req.Offset = i
			} else {
				errors.WriteError(w, r, errors.NewValidationError("Invalid query", map[string]string{"offset": "must be integer"}))
				return
			}
		}
		if v := chi.URLParam(r, "postID"); v != "" {
			req.PostID = v
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.ListComments(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		var out any
		out = resp
		data, err := json.Marshal(out)
		if err != nil {
			errors.WriteError(w, r, errors.New(http.StatusInternalServerError, "Internal Server Error", "Failed to encode response"))
			return
		}
		etag := fmt.Sprintf("W/\"%x\"", sha256.Sum256(data))
		w.Header().Set("ETag", etag)
		if match := r.Header.Get("If-None-Match"); match == etag {
			w.WriteHeader(http.StatusNotModified)
			return
		}
		_, _ = w.Write(data)

	}
}

// handlerCreateComment
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerCreateComment(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.CreateCommentRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if v := chi.URLParam(r, "postID"); v != "" {
			req.PostID = v
		}
		if req.UserID == "" {
			req.UserID = CurrentUserID(r)
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.CreateComment(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerGetPost
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerGetPost(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.GetPostRequest
		if v := r.URL.Query().Get("slug"); v != "" {
			req.Slug = v
		}
		if v := r.URL.Query().Get("limit"); v != "" {
			if i, err := strconv.Atoi(v); err == nil {
				req.Limit = i
			} else {
				errors.WriteError(w, r, errors.NewValidationError("Invalid query", map[string]string{"limit": "must be integer"}))
				return
			}
		}
		if v := r.URL.Query().Get("offset"); v != "" {
			if i, err := strconv.Atoi(v); err == nil {
				req.Offset = i
			} else {
				errors.WriteError(w, r, errors.NewValidationError("Invalid query", map[string]string{"offset": "must be integer"}))
				return
			}
		}
		if v := chi.URLParam(r, "slug"); v != "" {
			req.Slug = v
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.GetPost(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		var out any
		out = resp
		data, err := json.Marshal(out)
		if err != nil {
			errors.WriteError(w, r, errors.New(http.StatusInternalServerError, "Internal Server Error", "Failed to encode response"))
			return
		}
		etag := fmt.Sprintf("W/\"%x\"", sha256.Sum256(data))
		w.Header().Set("ETag", etag)
		if match := r.Header.Get("If-None-Match"); match == etag {
			w.WriteHeader(http.StatusNotModified)
			return
		}
		_, _ = w.Write(data)

	}
}

// handlerListTags
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerListTags(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.ListTagsRequest
		if v := r.URL.Query().Get("limit"); v != "" {
			if i, err := strconv.Atoi(v); err == nil {
				req.Limit = i
			} else {
				errors.WriteError(w, r, errors.NewValidationError("Invalid query", map[string]string{"limit": "must be integer"}))
				return
			}
		}
		if v := r.URL.Query().Get("offset"); v != "" {
			if i, err := strconv.Atoi(v); err == nil {
				req.Offset = i
			} else {
				errors.WriteError(w, r, errors.NewValidationError("Invalid query", map[string]string{"offset": "must be integer"}))
				return
			}
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.ListTags(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		var out any
		out = resp
		data, err := json.Marshal(out)
		if err != nil {
			errors.WriteError(w, r, errors.New(http.StatusInternalServerError, "Internal Server Error", "Failed to encode response"))
			return
		}
		etag := fmt.Sprintf("W/\"%x\"", sha256.Sum256(data))
		w.Header().Set("ETag", etag)
		if match := r.Header.Get("If-None-Match"); match == etag {
			w.WriteHeader(http.StatusNotModified)
			return
		}
		_, _ = w.Write(data)

	}
}

// handlerCreateTag
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerCreateTag(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.CreateTagRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.CreateTag(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerDeleteTag
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerDeleteTag(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.DeleteTagRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if v := chi.URLParam(r, "id"); v != "" {
			req.ID = v
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.DeleteTag(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}

// handlerUpdateTag
// Source: /home/strog/work/ang/cue/schema/types.cue:1252
func handlerUpdateTag(svc port.Blog) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var req port.UpdateTagRequest
		if err := decodeJSONRequest(r, &req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError("Invalid JSON", map[string]string{"body": err.Error()}))
			return
		}
		if v := chi.URLParam(r, "id"); v != "" {
			req.ID = v
		}
		if err := validate.Struct(&req); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		if err := req.Validate(); err != nil {
			errors.WriteError(w, r, errors.NewValidationError(err.Error(), nil))
			return
		}
		resp, err := svc.UpdateTag(r.Context(), req)
		if err != nil {
			errors.WriteError(w, r, err)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)

	}
}