# python/api/openapi.yaml
openapi: 3.0.3
info:
  title: ANG API
  version: 0.1.0
  x-ang-version: 0.1.127
  x-ang-input-hash: 7d9ba8054e7ed584000a9c7ea12959fe1c66741726ec018e6c1e6f79ce78470c
  x-ang-compiler-hash: 294b50c17c5868adf98ff6e7e34d9b0ddcf67156d863f2f508ad57df44889164
paths:
...
  /comments/{id}:
    put:
      summary: UpdateComment
      operationId: UpdateComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - bearerAuth: []
  /my/posts:
    get:
      summary: ListMyPosts
      operationId: ListMyPosts
      parameters:
        - in: query
          name: userid
          required: true
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            maximum: 100
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
      x-auth-roles:
        - author
        - admin
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListMyPostsResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - bearerAuth: []
...
  /posts:
    post:
      summary: CreatePost
      operationId: CreatePost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      x-auth-roles:
        - author
        - admin
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - bearerAuth: []
...
  /posts/{id}:
    put:
      summary: UpdatePost
      operationId: UpdatePost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      x-auth-roles:
        - author
        - admin
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - bearerAuth: []
  /posts/{id}/archive:
    post:
      summary: ArchivePost
      operationId: ArchivePost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchivePostRequest'
      x-auth-roles:
        - author
        - admin
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchivePostResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - bearerAuth: []
  /posts/{id}/publish:
    post:
      summary: PublishPost
      operationId: PublishPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishPostRequest'
      x-auth-roles:
        - admin
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishPostResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - bearerAuth: []
  /posts/{id}/submit:
    post:
      summary: SubmitPost
      operationId: SubmitPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitPostRequest'
      x-auth-roles:
        - author
        - admin
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitPostResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - bearerAuth: []
...
  /posts/{postID}/comments:
    post:
      summary: CreateComment
      operationId: CreateComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - bearerAuth: []
  /posts/{slug}:
    get:
      summary: GetPost
      operationId: GetPost
      parameters:
        - in: query
          name: slug
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            maximum: 100
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPostResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /tags:
    post:
      summary: CreateTag
      operationId: CreateTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      x-auth-roles:
        - admin
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - bearerAuth: []
...
  /tags/{id}:
    put:
      summary: UpdateTag
      operationId: UpdateTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagRequest'
      x-auth-roles:
        - admin
      responses:
        '200':
          description: OK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagResponse'
        '401':
          description: Unauthorized
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - bearerAuth: []
...
components:
  schemas:
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: integer
        errors:
          type: object
    ArchivePostRequest:
      type: object
      properties:
        id:
          type: string
      required:
        - id
    ArchivePostResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
    CreateCommentRequest:
      type: object
      properties:
        postid:
          type: string
        content:
          type: string
          minLength: 1
          maxLength: 1000
        userid:
          type: string
      required:
        - postid
        - content
        - userid
    CreateCommentResponse:
      type: object
      properties:
        id:
          type: string
      required:
        - id
    CreatePostRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 200
        content:
          type: string
        userid:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - content
        - userid
    CreatePostResponse:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
      required:
        - id
        - slug
    CreateTagRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        description:
          type: string
      required:
        - name
    CreateTagResponse:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
      required:
        - id
        - slug
...
    GetPostResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        authorid:
          type: string
        createdat:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - content
        - authorid
        - createdat
        - tags
...
    ListMyPostsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: string
      required:
        - data
...
    PublishPostRequest:
      type: object
      properties:
        id:
          type: string
      required:
        - id
    PublishPostResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
    SubmitPostRequest:
      type: object
      properties:
        id:
          type: string
      required:
        - id
    SubmitPostResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    UpdateCommentRequest:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
          minLength: 1
          maxLength: 1000
        userid:
          type: string
      required:
        - id
        - content
        - userid
    UpdateCommentResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    UpdatePostRequest:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          minLength: 5
          maxLength: 200
        content:
          type: string
      required:
        - id
    UpdatePostResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
    UpdateTagResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT